Moving performance counters away
