+++ src/MassTransit.Host/MassTransit.Host.csproj
@@ -41,5 +41,6 @@
     <ProjectReference Include="..\Containers\MassTransit.AutofacIntegration\MassTransit.AutofacIntegration.csproj" />
     <ProjectReference Include="..\Loggers\MassTransit.Log4NetIntegration\MassTransit.Log4NetIntegration.csproj" />
     <ProjectReference Include="..\MassTransit\MassTransit.csproj" />
+    <ProjectReference Include="..\Monitoring\MassTransit.Monitoring.Windows\MassTransit.Monitoring.Windows.csproj" />
   </ItemGroup>
 </Project>
