--- src/MassTransit/Monitoring/Performance/StatsD/StatsDCounterFactory.cs
@@ -1,29 +1,0 @@
-﻿// Copyright 2007-2016 Chris Patterson, Dru Sellers, Travis Smith, et. al.
-//  
-// Licensed under the Apache License, Version 2.0 (the "License"); you may not use
-// this file except in compliance with the License. You may obtain a copy of the 
-// License at 
-// 
-//     http://www.apache.org/licenses/LICENSE-2.0 
-// 
-// Unless required by applicable law or agreed to in writing, software distributed
-// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
-// CONDITIONS OF ANY KIND, either express or implied. See the License for the 
-// specific language governing permissions and limitations under the License.
-namespace MassTransit.Monitoring.Performance.StatsD
-{
-    public class StatsDCounterFactory : ICounterFactory
-    {
-        readonly StatsDConfiguration _config;
-
-        public StatsDCounterFactory(StatsDConfiguration config)
-        {
-            _config = config;
-        }
-
-        public IPerformanceCounter Create(CounterCategory category, string counterName, string instanceName)
-        {
-            return new StatsDPerformanceCounter(_config, category.Name, counterName, instanceName );
-        }
-    }
-}
