+++ src/Monitoring/MassTransit.Monitoring.Windows/Performance/ConsumerPerformanceCounterCache.cs
@@ -1,0 +1,44 @@
+﻿// Copyright 2007-2015 Chris Patterson, Dru Sellers, Travis Smith, et. al.
+//  
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+// this file except in compliance with the License. You may obtain a copy of the 
+// License at 
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0 
+// 
+// Unless required by applicable law or agreed to in writing, software distributed
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the 
+// specific language governing permissions and limitations under the License.
+namespace MassTransit.Monitoring.Performance
+{
+    using System;
+    using System.Collections.Concurrent;
+    using Windows;
+
+
+    public class ConsumerPerformanceCounterCache
+    {
+        readonly ConcurrentDictionary<string, Lazy<ConsumerPerformanceCounter>> _types;
+
+        ConsumerPerformanceCounterCache()
+        {
+            _types = new ConcurrentDictionary<string, Lazy<ConsumerPerformanceCounter>>();
+        }
+
+        public static IConsumerPerformanceCounter GetCounter(ICounterFactory factory, string consumerType)
+        {
+            return Cached.Counter.Value._types
+                .GetOrAdd(consumerType, x => new Lazy<ConsumerPerformanceCounter>(() => new ConsumerPerformanceCounter(factory, x)))
+                .Value;
+        }
+
+
+        static class Cached
+        {
+
+            public static readonly Lazy<ConsumerPerformanceCounterCache> Counter =
+                new Lazy<ConsumerPerformanceCounterCache>(() => new ConsumerPerformanceCounterCache());
+        }
+    }
+}
